Added image upload endpoint added role controller and manager added sweetalert

+++ Angular/src/app/core/components/administration/administration.component.ts
@@ -1,18 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from '../../models/User';
 import { UserService } from '../../services/UserService';
+import Swal from 'sweetalert2';
 
 @Component({
-  selector: 'app-users',
-  templateUrl: './users.component.html',
-  styleUrls: ['./users.component.less']
+  selector: 'app-administration',
+  templateUrl: './administration.component.html',
+  styleUrls: ['./administration.component.less']
 })
 export class AdministrationComponent implements OnInit {
   users: User[] = [];
-  roles: string[] = ['Manager', 'Employee'];
-  projects: string[] = ['Project A', 'Project B', 'Project C'];
-  selectedUser: User = new User();
-  selectedProject: number = 0;
 
   constructor(private userService: UserService) { }
 
@@ -22,9 +19,19 @@ export class AdministrationComponent implements OnInit {
 
 
   deleteUser(user: User) {
-    this.userService.deleteUser(user).subscribe(() => {
-      const index = this.users.findIndex(u => u.id === user.id);
-      this.users.splice(index, 1);
-    });
+    Swal.fire({
+      title: 'Are you sure?',
+      icon: 'info',
+      confirmButtonText: 'Delete'
+    }).then((result) => {
+      if (result['isConfirmed']){
+        this.userService.deleteUser(user).subscribe(() => {
+         
+          const index = this.users.findIndex(u => u.id === user.id);
+          this.users.splice(index, 1);
+        });
+      }
+    })
+ 
   }
 }
